Move get_criterion and get_optimizer functions to train script
